ScrapeFunctionApp/scrape_trigger/requirements.txt
azure-functions
azure-storage-blob
beautifulsoup4
requests
